fix: addalert code splitting part 4
